+++ src/cli/zithc-main.cpp
@@ -1,6 +1,5 @@
 #include "cli/compiler-driver.hpp"
 #include "cli/options.hpp"
-#include "cli/zith-flags.hpp"
 
 #include <cstdio>
 #include <cstdlib>
@@ -9,6 +8,5 @@
 
 int main(int argc, char **argv) {
     auto opts = zith::cli::parseArgs(argc, argv);
-    auto defaults = zith::cli::loadZithFlags();
-    zith::cli::CompilerDriver::run(opts);
+    return zith::cli::CompilerDriver::run(opts);
 }